+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Changelog
 
 ## [1.1.2] - 2025-12-03
-**1** commit in this release
+**2** commits in this release
 
 
 
-### 🐛 Bug Fixes
+### ✨ Features
 
-- Use force flag when syncing tags to public repo (5da93be)
+- Implement conditional debug logging (227e8b7)
+
+### 📝 Other Changes
+
+- Karthikeyan R ()